Blockchain persistence + merge main

+++ backend/pong/game.py
@@ -1,19 +1,28 @@
 from django.core.exceptions import ObjectDoesNotExist
 import json
 from asgiref.sync import sync_to_async
-from channels.generic.websocket import AsyncWebsocketConsumer
+# from channels.generic.websocket import AsyncWebsocketConsumer
 from game.models import RoomsModel, PlayerRoomModel, PlayersModel
 
-import asyncio
+from django.core.exceptions import MultipleObjectsReturned, ObjectDoesNotExist
+
+# import asyncio
 
 from .data import pong_data
 import random
 
 @sync_to_async
 def get_info(consumer):
-    consumer.room = RoomsModel.objects.get(id=consumer.room_id)
-    consumer.player = PlayerRoomModel.objects.get(id=consumer.player_id)
-    consumer.server = PlayerRoomModel.objects.get(player=consumer.room.server)
+    try:
+        consumer.room = RoomsModel.objects.get(id=consumer.room_id)
+        consumer.player = PlayerRoomModel.objects.get(id=consumer.player_id)
+        consumer.server = PlayerRoomModel.objects.get(player=consumer.room.server)
+    except ObjectDoesNotExist:
+        return False
+    except MultipleObjectsReturned:
+        return False
+    return True
+    
 
 @sync_to_async
 def get_room_data(players, room_id):
@@ -24,11 +33,15 @@ def get_room_data(players, room_id):
             'players': [{'x': i.x, 'y': i.y} for i in players]
         })
     except ObjectDoesNotExist:
-        return "Error: Rooms not found"
+        print(f"Room with ID {room_id} does not exist.")
 
 @sync_to_async
 def get_teams_data(consumer, room_id):
-    consumer.room = RoomsModel.objects.get(id=consumer.room_id)
+    try:
+        consumer.room = RoomsModel.objects.get(id=consumer.room_id)
+    except RoomsModel.DoesNotExist:
+        print(f"Room with ID {room_id} does not exist.")
+        return
     consumer.player = PlayerRoomModel.objects.get(id=consumer.player_id)
     consumer.server = PlayerRoomModel.objects.get(player=consumer.room.server)
     players0 = PlayerRoomModel.objects.filter(room=room_id, side=0)
@@ -42,6 +55,22 @@ def get_teams_data(consumer, room_id):
 def get_score_data(room_id):
     room = RoomsModel.objects.get(id=room_id)
     return json.dumps({ 'score': [room.score0, room.score1] })
+
+@sync_to_async
+def get_win_data(room_id):
+    room = RoomsModel.objects.get(id=room_id)
+    if room.score0 > room.score1:
+        winner = 'player0'
+        winning_score = room.score0
+    elif room.score1 > room.score0:
+        winner = 'player1'
+        winning_score = room.score1
+    return json.dumps({
+        'win': winner,
+        'score': [room.score0, room.score1],
+        'winning_score': winning_score,
+        'roomid': room_id
+    })
 
 @sync_to_async
 def start_game(consumer):
@@ -70,14 +99,34 @@ def end_game(consumer):
 
 @sync_to_async
 def quit(consumer):
+    if PlayerRoomModel.objects.filter(room=consumer.room_id).count() == 0:
+        return
     if PlayerRoomModel.objects.filter(room=consumer.room_id).count() == 1:
         consumer.room.delete()
+        return
+    if consumer.player == None:
         return
     if consumer.server == consumer.player:
         consumer.player.delete()
         change_server(consumer, PlayerRoomModel.objects.filter(room=consumer.room_id).first())
     else:
         consumer.player.delete()
+
+@sync_to_async
+def remove_player(consumer):
+    if consumer.player is not None:
+        consumer.player.delete()
+
+@sync_to_async
+def check_player(consumer):
+    try:
+        consumer.player = PlayerRoomModel.objects.get(id=consumer.player_id)
+    except PlayerRoomModel.DoesNotExist:
+        print(f"Player with ID {consumer.player_id} does not exist in game.")
+        return False
+    if (consumer.player == None):
+        return False
+    return True
 
 @sync_to_async
 def change_server_direction(consumer):
